Add a mask layer that is hidden by default from the user unless asked for.
This is to allow for data being 'decayed out' so a filled cell goes back to nan.
This case was failing otherwise as the nan was being interpreted as 'no change' rather than 'make it empty' in the diffs.
This could be avoided by holding full arrays at each survey commit and would be faster, is the storage gain worth the cost?

working on xipe2